To rename server package to threadpool package.
